tests/test_project/pyproject.toml
[project]
name = "test-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-boot[db]",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "tortoise-orm==0.24.1",
]

[tool.uv.sources]
fastapi-boot = { workspace = true }
